requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
discord==2.3.2
discord.py==2.3.2
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0
duplicity==0.8.21
fasteners==0.14.1
frozenlist==1.4.1
future==0.18.2
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mutagen==1.47.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.19.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
usb-creator==0.3.7
wadllib==1.3.6
websockets==12.0
xdg==5
xkit==0.0.0
yarl==1.9.4
youtube_dl @ git+https://github.com/ytdl-org/youtube-dl.git@be008e657d79832642e2158557c899249c9e31cd
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz#sha256=8f8f6748886903c07090ff6901b8f82b5f79e39721d4d0a530274262a48d4693
zipp==1.0.0

youtube_dl~=2021.12.17